# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "향수 정보 서비스 A.fume Server Api 문서"
  version: 1.0.0
  title: "Swagger A.fume Server"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "heesung6701@naver.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "{SERVER_URL}:{SERVER_PORT}"
basePath: "/A.fume/api/0.0.1"
...
schemes:
#- "https"
- "http"
# security:
# - APIKeyHeader: []
paths:
...
  /perfume/{perfumeIdx}/review:
    post:
      tags:
      - "review"
      summary: 시향노트 추가"
      description: "특정 향수에 시향노트 추가하기"
      operationId: "postReview"
      security:
        - userToken: []
      x-security-scopes:
        - user
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "perfumeIdx"
        in: "path"
        description: "향수 Idx"
        required: true
        type: "integer"
        format: "int64"
      - name: "body"
        in: "body"
        description: "시향노트 정보"
        required: true
        schema:
          $ref: "#/definitions/ReviewInfo"
      responses:
        "200":
          description: "성공"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "Review"
    get:
      tags:
      - "review"
      summary: "전체 시향노트 반환(별점순)"
      description: "특정 향수에 달린 전체 시향노트 별점순으로 가져오기"
      operationId: "getReviewOfPerfumeByScore"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "perfumeIdx"
        in: "path"
        description: "향수 Idx"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              allOf:
              - $ref: "#/definitions/ReviewInfo"
              - type: "object"
                properties:
                  reviewIdx:
                    type: "integer"
                    format: "int64"
                example:
                  reviewIdx: 1
        "400":
          description: "Invalid username/password supplied"
      x-swagger-router-controller: "Review"
  /perfume/{perfumeIdx}/review/recent:
    get:
      tags:
      - "review"
      summary: "전체 시향노트 반환(최신순)"
      description: "특정 향수에 달린 전체 시향노트 최신순으로 가져오기"
      operationId: "getReviewOfPerfumeByRecent"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "perfumeIdx"
        in: "path"
        description: "향수 Idx"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              allOf:
              - $ref: "#/definitions/ReviewInfo"
              - type: "object"
                properties:
                  reviewIdx:
                    type: "integer"
                    format: "int64"
                example:
                  reviewIdx: 1
        "400":
          description: "Invalid username/password supplied"
      x-swagger-router-controller: "Review"
  /review/{reviewIdx}:
    get:
      tags:
      - "review"
      summary: "시향노트 반환"
      description: "특정 시향노트 가져오기"
      operationId: "getReviewByIdx"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "reviewIdx"
        in: "path"
        description: "시향노트 Idx"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
          schema:
            $ref: "#/definitions/ReviewInfo"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
        "404":
          description: "Review not found"
      x-swagger-router-controller: "Review"
    put:
      tags:
      - "review"
      summary: "시향노트 수정"
      description: "시향노트 수정하기"
      operationId: "updateReview"
      security:
        - userToken: []
      x-security-scopes:
        - user
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "reviewIdx"
        in: "path"
        description: "시향노트 Idx"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/ReviewInfo"
      responses:
        "200":
          description: "성공"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "Review"
    delete:
      tags:
      - "review"
      summary: "시향노트 삭제"
      description: "시향노트 삭제하기"
      operationId: "deleteReview"
      security:
        - userToken: []
      x-security-scopes:
        - user
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "reviewIdx"
        in: "path"
        description: "시향노트 Idx"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      x-swagger-router-controller: "Review"
...
  /user:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: "This can only be done by the logged in user."
      operationId: "createUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          allOf:
          - $ref: "#/definitions/User"
          - type: "object"
            properties:
              password:
                type: "string"
            example:
              password: "test"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "User"
  /user/login:
    get:
      tags:
      - "user"
      summary: "Logs user into the system"
      operationId: "loginUser"
      produces:
      - "application/json"
      parameters:
      - name: "email"
        in: "query"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: success
          schema:
            type: object
            properties:
              message:
                type: string
                example: 로그인 성공
              data:
                type: object
                properties:
                  token:
                    type: string
                    description: login용 userToken
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWR4IjoxLCJuaWNrbmFtZSI6IuycpO2drOyEsSIsImdlbmRlciI6Im1hbGUiLCJwaG9uZSI6IjAxMC0yMDgxLTM4IiwiZW1haWwiOiJoZWUueW91bkBzYW1zdW5nLmNvbSIsImJpcnRoIjoxOTk1LCJpYXQiOjE2MDcwMDU3NTIsImV4cCI6MTYwODczMzc1MiwiaXNzIjoiYWZ1bWUtamFja3BvdCJ9.wHBPi8pQuzJRbUy4noAwdwhCRtA9mG_UrI-S0IUEkxY
                  refreshToken:
                    type: string
                    description: token 재발급 용 
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyZWZyZXNoVG9rZW4iOnsidXNlcklkeCI6MSwibmlja25hbWUiOiLsnKTtnazshLEiLCJnZW5kZXIiOiJtYWxlIiwicGhvbmUiOiIwMTAtMjA4MS0zOCIsImVtYWlsIjoiaGVlLnlvdW5Ac2Ftc3VuZy5jb20iLCJiaXJ0aCI6MTk5NX0sImlhdCI6MTYwNzAwNTc1MiwiZXhwIjoxNjA4NzMzNzUyLCJpc3MiOiJhZnVtZS1qYWNrcG90In0.ePRhCLytxo-FlKLsxE84TvPee8G3C4LbsdUNBQ-OcF8
        "400":
          description: "Invalid username/password supplied"
      x-swagger-router-controller: "User"
...
  /user/{userIdx}:
    get:
      tags:
      - "user"
      summary: "Get user by user idx"
      operationId: "getUserByIdx"
      produces:
      - "application/json"
      parameters:
      - name: "userIdx"
        in: "path"
        description: "유저 ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            allOf:
            - $ref: "#/definitions/User"
            - type: "object"
              properties:
                userIdx:
                  type: "integer"
              example:
                userIdx: 1
        "400":
          description: "Invalid userIdx supplied"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
    put:
      tags:
      - "user"
      summary: "Updated user"
      description: "This can only be done by the logged in user."
      operationId: "updateUser"
      produces:
      - "application/json"
      parameters:
      - name: "userIdx"
        in: "path"
        description: "name that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          allOf:
          - $ref: "#/definitions/User"
          - type: "object"
            properties:
              userIdx:
                type: "integer"
            example:
              userIdx: 1
      responses:
        "400":
          description: "Invalid user supplied"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
    delete:
      tags:
      - "user"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "userIdx"
        in: "path"
        description: "The Idx that needs to be deleted"
        required: true
        type: "string"
      responses:
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
...
securityDefinitions:
  userToken:
      type: apiKey
      name: x-access-token
      in: header
      description: /Bearer {token}/ 를 입력해주시면 됩니다.
  APIKeyHeader:
    type: apiKey
    name: X-API-Key
    in: header
definitions:
  User:
    type: "object"
    properties:
      email:
        type: "string"
      nickname:
        type: "string"
      gender:
        type: "string"
      phone:
        type: "string"
      birth:
        type: "string"
    example:
      email: "hee.youn@samsung.com"
      nickname: "쿼카맨"
      gender: "남자"
      phone: "010-2081-3818"
      birth: "1995"
...
  ReviewInfo:
    type: "object"
    properties:
      userIdx:
        description: "회원 고유 번호"
        type: "integer"
        format: "int64"
      score:
        description: "평점"
        type: "number"
        minimum: 0
        maximum: 5
      content:
        description: "내용"
        type: "string"  
      longevity:
        description: "지속력"
        type: "string"
        enum:
        - "매우 약함"
        - "약함"
        - "보통"
        - "강함"
        - "매우 강함"
      sillage:
        description: "잔향감"
        type: "string"
        enum:
        - "가벼움"
        - "보통"
        - "무거움"
      seasonal:
        description: "계절감"
        type: "array"
        items:
          type: "string"
          enum:
            - "봄"
            - "여름"
            - "가을"
            - "겨울"        
      gender:
        description: "성별감"
        type: "string"
        enum:
        - "남성"
        - "중성"
        - "여성"
      access: 
        description: "시향노트 공유 여부"
        type: "string"
        enum:
        - "true"
        - "false"
    example:
      userIdx: 0
      score: 2
      longevity: "강함"
      sillage: "보통"
      seasonal:
      - "겨울"
      - "가을"
      gender: "중성"
      access: "true"
      content: "시향노트 내용"
...